+++ lib/magic/presenter.rb
@@ -9,11 +9,13 @@ module Magic # :nodoc:
 	# Presentation layer for Rails models
 	module Presenter
 		autoload :Base,      'magic/presenter/base'
+		autoload :Helpers,   'magic/presenter/helpers'
 		autoload :GlobalID,  'magic/presenter/global_id'
 		autoload :Generator, 'generators/magic/presenter/generator'
 
 		singleton_class.delegate *%i[
 				for name_for
+				view_context view_context=
 		], to: Base
 	end