+++ spec/internal/app/controllers/application_controller.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+class ApplicationController < ActionController::Base
+	private
+
+	def request
+		super or
+				ActionDispatch::TestRequest.create
+	end
+end