+++ spec/internal/app/presenters/person_presenter.rb
@@ -2,4 +2,6 @@
 
 class PersonPresenter < ApplicationPresenter
 	def name = "#{first_name} #{last_name}"
+
+	def to_s = name
 end